Read environment file in index.js
I think the EnvironmentFile in a service is not affected by its
RootDirectory, so it doesn’t work in a portable service. Let’s read .env
in the index.js instead. (It’s still possible to inherit the values from
the environment instead, though I’m not sure how useful that is.)