Make non-bio* biwords more likely
bio* is a common prefix, but the results from it tend to be not overly
exciting, in my opinion. Let’s give other words a bit more of a chance.

This could be done in the application logic, of course, but repeating
the other words in the word list itself seems like the easiest approach.
Done with the following command (GNU sed only):

    sed -i '/^.\?bio/I!p' biwords