rust: Initial commit
I’m starting a Rust version of the bot. The primary motivation is that I
just want to play around with Rust; the primary justification for
following that urge is that it should shrink the deployment image
massively from the current over half a gigabyte.

This initial version only prints one random biword to stdout.