rust: Keep surrounding text in replacement()
Just replace the “bi” part of the biword and keep the rest. The JS
version does this with regular expressions, but here we use a kind of
linear state machine instead, manually iterating over the (decomposed)
characters. The vowel flag isn’t used yet, which shows some warnings,
but that’ll be taken care of separately.