rust: Randomly replace with “straight” or “hetero”
The randomness is injected, so that in the tests we can inject a fake
RNG and expect an exact value. A test for the static_bool_rng() helper
is provided; however, given that rand’s output is supposed to be stable
given the same inputs, the test is ignored by default.