rust: Move binary crate to bin/
This is going to be the analog of the split between tweet.js and
index.js – one binary to print a tweet text to stdout, and one to
actually send a tweet. (But I’m not making the mistake again of calling
the binary that *doesn’t* send a tweet “tweet”, that’s stupid. “text”
isn’t great but it’s something.)