rust: Add tweet binary
Using the dotenv library [1] to load the credentials from the .env file,
just like the JS version, and the rust_twitter_bot_lib library [2] to
send the tweet.

This pulls in *a lot* of dependencies, and I suspect they’re coming from
rust_twitter_bot_lib, not dotenv. Maybe I can try a different library…
but for now, this works, so let’s commit it.

[1]: https://docs.rs/dotenv/0.15.0/dotenv/
[2]: https://docs.rs/rust_twitter_bot_lib/0.1.2/rust_twitter_bot_lib/