rust: Make image world-readable
Otherwise, the service can’t change into its working directory and fails
to start. Unfortunately, this leaves the .env in the deployment image
world-readable and -searchable on the development system, but the JS
version has the same problem, so I don’t feel obliged to fix it in this
commit (though it should be fixed at some point).