Make deployment images not world searchable
Use a dual approach to protect the world-readable .env file within the
image: on the development system, make the image itself not
world-searchable; on the deployment system, all of /var/lib/portables/
isn’t world-searchable and the deployment image can be made
world-searchable inside there so that the service can use it.

I tested both the Rust and JS versions with this and both work now
without having a world-accessible .env anywhere.