Remove Birkenhead drill from biwords
The concept or practice of taking women and children first in an
evacuation or rescue, according to enwiktionary [1]; named after the
maritime disaster of HMS Birkenhead. The enwiki article [2] sounds like
the term could also carry a somewhat positive connotation of courage and
honour, but on the whole I think it’s better to remove it. (But I’m
keeping the extant post [3][4] up.)

[1]: https://en.wiktionary.org/wiki/Birkenhead_drill
[2]: https://en.wikipedia.org/wiki/HMS_Birkenhead_(1845)#Legacy
[3]: https://botsin.space/@ItsBiNotStraight/109661597424799199
[4]: https://twitter.com/ItsBiNotHetero/status/1612576546614644736