Remove -phob* biwords
The bot posted biophobic [1], which seems to be mostly harmless in
meaning (“having a dislike of nature”), but it’s still not great to keep
in the word list either. More importantly, this reminded me to remove
all words related to biphobia, which we definitely don’t want;
bibliophobia is also not nice, kick that out too.

I also checked the -phil* biwords, in case there were any more monsters
like biastophilia lurking around (see 8dfc0a5f31), but all of the
remaining ones seem fine.

[1]: https://botsin.space/@ItsBiNotStraight/112226524065873607