Move from botsin.space to fox.nexus
botsin.space is shutting down (thank you for several years of service ❤️),
so move to another friendly instance.

Leave botsin.space in the .env.template, because it seems like a
reasonable example instance name even after the instance is no more.
(I might have to change it if the domain ever lapses. We’ll see.)